feat: add test for create embedded
