+++ lib/resolvers/index.ts
@@ -1,0 +1,3 @@
+export * from './crud.resolver';
+export * from './embedded.resolver';
+export * from './references-many.resolver';
